separate interfaces and defaults of loadbalancer and stream
