+++ proxy/loadbalancer/defaults.go
@@ -1,4 +1,4 @@
-package proxy
+package loadbalancer
 
 import (
 	"m3u-stream-merger/store"
