+++ proxy/stream/interfaces.go
@@ -1,0 +1,13 @@
+package stream
+
+import (
+	"net/http"
+)
+
+type ResponseWriter interface {
+	http.ResponseWriter
+}
+
+type StreamFlusher interface {
+	Flush()
+}
